Continuing through Chapter 10
Adds pagination to users index page.

A few minor .css changes to keep the footer at the bottom of the
scrollable index page.

Adds a test for user index pagination.

Note:  There is some depreciation warning in the will_paginate gem
that's not a huge problem now but I would like to fix in the future.